DOC: Better documentation of realign_raw (#12135)

+++ mne/surface.py
@@ -772,7 +772,7 @@ class _CheckInside:
 
 def _fread3(fobj):
     """Read 3 bytes and adjust."""
-    b1, b2, b3 = np.fromfile(fobj, ">u1", 3)
+    b1, b2, b3 = np.fromfile(fobj, ">u1", 3).astype(np.int64)
     return (b1 << 16) + (b2 << 8) + b3
